refactor: 패키지 구조 나눔
