--- src/main/java/com/example/e2ekernelengine/crawler/StringToTimestamp.java
@@ -1,5 +1,0 @@
-package com.example.e2ekernelengine.crawler;
-
-public class StringToTimestamp {
-
-}
