+++ src/main/java/com/example/e2ekernelengine/crawler/controller/CrawlerController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.e2ekernelengine.crawler.ChannelRssCrawler;
 import com.example.e2ekernelengine.crawler.dto.testRssBlogCrawlerDto;
+import com.example.e2ekernelengine.crawler.service.ChannelRssCrawler;
 
 import lombok.RequiredArgsConstructor;
