+++ src/main/java/com/example/e2ekernelengine/crawler/service/AbstractRssCrawler.java
@@ -1,4 +1,4 @@
-package com.example.e2ekernelengine.crawler;
+package com.example.e2ekernelengine.crawler.service;
 
 import java.io.IOException;
 import java.util.regex.Matcher;
