+++ src/main/java/com/example/e2ekernelengine/crawler/service/IRssCrawler.java
@@ -1,7 +1,9 @@
-package com.example.e2ekernelengine.crawler;
+package com.example.e2ekernelengine.crawler.service;
 
 import java.sql.Timestamp;
 import java.util.List;
+
+import com.example.e2ekernelengine.crawler.dto.FeedDataDto;
 
 public interface IRssCrawler {
